tests: add insider bypass test
Interfaces are easily to prove to be completely bypassed by checking their
function header. But to prove insiders are bypassed too, we need to insert
an invalid instruction at the original function (of the vmlinux).

After we've done it, we can run some coverage tests. Reusing existing
those existing tests is a good idea. For now CPU Hotplug, CFS BWC functions
and scheduler syscalls functions are already covered.

Signed-off-by: Yihao Wu <wuyihao@linux.alibaba.com>